requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
Wavelink==2.0.3
yt_dlp==2023.3.4
